Add running of entire e2e data set
