+++ .test.env
@@ -8,7 +8,6 @@ SMPC__DATABASE__MIGRATE=true
 SMPC__DATABASE__CREATE=true
 SMPC__DATABASE__LOAD_PARALLELISM=8
 SMPC__MAX_BATCH_SIZE=64
-SMPC__INIT_DB_SIZE=1000
 SMPC__MAX_DB_SIZE=10000
 SMPC__HAWK_REQUEST_PARALLELISM=10
 SMPC__AWS__REGION=us-east-1
