+++ iris-mpc/bin/client/client.rs
@@ -1,5 +1,5 @@
 use clap::Parser;
-use iris_mpc::client::{run_client, Opt};
+use iris_mpc::client::e2e::{run_client, Opt};
 
 #[tokio::main]
 async fn main() -> eyre::Result<()> {
