+++ scripts/run-client.sh
@@ -10,5 +10,4 @@ cargo run --bin client -- \
     --request-topic-arn arn:aws:sns:us-east-1:000000000000:iris-mpc-input.fifo \
     --requests-bucket-name wf-smpcv2-dev-sns-requests \
     --public-key-base-url "http://localhost:4566/wf-dev-public-keys" \
-    --region us-east-1 \
     --response-queue-url http://sqs.us-east-1.localhost.localstack.cloud:4566/000000000000/iris-mpc-results-us-east-1.fifo
